mcs-examples/pyproject.toml
# mcs-examples/pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "mcs-examples"
version         = "0.1.1"
description     = "Minimal MCS client examples"
requires-python = ">=3.9"
dependencies    = [
  "mcs-driver-core>=0.2",
  "requests>=2.32",
  "litellm>=1.73",
  "python-dotenv>=1.1.0"
]

[project.optional-dependencies]
quickstart = [
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "pydantic>=2.7",
]
# Damit im Beispiel-Code die Treiber-Pakete verfügbar sind,
# installiere sie lokal per editable install:
# pip install -e ..   (installs mcs-driver-core from the repo root)
# pip install -e ../mcs-driver-rest-http
#
# Hier müssen keine runtime-Dependencies eingetragen werden,
# denn die Treiber-Pakete liefern ihre eigenen Abhängigkeiten.

[tool.setuptools.packages.find]
where   = ["."]
include = ["*"]
